Modified project layout to conform to setuptools
- No code change
- Modified test imports
- Included nosetests to run setuptools tests from setup.py
