+++ .gitignore
@@ -21,6 +21,7 @@ var/
 *.egg-info/
 .installed.cfg
 *.egg
+*.eggs/
 
 # Installer logs
 pip-log.txt
